Update BoardSections.php
No need to be logged in to get board sections. 
You don't need to be logged in to get a list of boards for any user so why would you need to be to get sections? I made a test where I got all the boards and all the sections of a random pinner and it worked.
It's just super annoying to log after you get all the boards just to view the sections.